Testing Fetching Movie Details and Save Them into The Database
Updated the functionality. Now I can search for a movie by title, and I get a result list. From the result list, I can click on a movie that I like. If the movie exists in the database, it is rendered from the database; otherwise, it is fetched using the APIs.However, the organization of the result list is not properly done yet. The CSS needs an adjustment.